long -> Long

+++ src/main/java/br/com/henriquesousa/musicapp/entity/User.java
@@ -2,18 +2,18 @@ package br.com.henriquesousa.musicapp.entity;
 
 public class User {
 
-    private long id;
+    private Long id;
     private String name;
 
-    public User(long id, String name) {
+    public User(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getName() {
